Restructured signaling_server install for better practices with script
