+++ opentera-webrtc-native-client/OpenteraWebrtcNativeClient/python/package/CMakeLists.txt
@@ -10,6 +10,8 @@ set(ENABLE_HTML_DOC ${OPENTERA_WEBRTC_ENABLE_PYTHON_HTML_DOC})
 set(ENABLE_HTML_DOC OFF)
 
 set(PYTHON_SUBPACKAGE_NAME "native_client")
+set(PYTHON_PACKAGE_DIR ${CMAKE_CURRENT_BINARY_DIR}/opentera_webrtc_${PYTHON_SUBPACKAGE_NAME})
+set(PYTHON_PACKAGE_CONTENT_DIR ${PYTHON_PACKAGE_DIR}/opentera/webrtc/${PYTHON_SUBPACKAGE_NAME})
 
 include(../../../../cmake/pip_utils.cmake)
