+++ signaling-server/opentera-signaling-server
@@ -1,0 +1,7 @@
+#!python
+
+from opentera.webrtc.signaling_server.signaling_server import main
+
+
+if __name__ == '__main__':
+    main()
